+++ packages/mobile/App/dummyData/referenceData.ts
@@ -4,12 +4,12 @@ import { SyncRecord } from '~/services/sync/source';
 
 // for dummy data generation
 import { generatePatient } from '~/dummyData/patients';
-import { ICD10_DIAGNOSES } from './diagnoses';
+import { DIAGNOSES } from './diagnoses';
 import { splitIds } from './utilities';
 
 const generator = new Chance('patients');
 const DUMMY_PATIENT_COUNT = 44;
-const dummyPatients = (new Array(DUMMY_PATIENT_COUNT))
+const dummyPatients = new Array(DUMMY_PATIENT_COUNT)
   .fill(0)
   .map(() => generatePatient(generator))
   .map(p => ({
@@ -17,10 +17,8 @@ const dummyPatients = (new Array(DUMMY_PATIENT_COUNT))
     lastModified: generator.date({ year: 1971, month: 0, day: 0 }),
   }));
 
-const sortByModified = (
-  a: SyncRecord,
-  b: SyncRecord,
-): any => a.data.lastModified - b.data.lastModified;
+const sortByModified = (a: SyncRecord, b: SyncRecord): any =>
+  a.data.lastModified - b.data.lastModified;
 
 const dummyPatientRecords: SyncRecord[] = dummyPatients.map(p => ({
   data: p,
@@ -34,7 +32,9 @@ const makeRefRecords = (referenceDataType: ReferenceDataType, values: string): I
     lastModified: generator.date({ year: 1971, month: 1, day: 0 }),
   }));
 
-const VILLAGES = makeRefRecords(ReferenceDataType.Village, `
+const villageRefData = makeRefRecords(
+  ReferenceDataType.Village,
+  `
   Ba
   Lami
   Levuka
@@ -46,20 +46,16 @@ const VILLAGES = makeRefRecords(ReferenceDataType.Village, `
   Navua
   Korovou
   Nasinu
-`);
+`,
+);
 
-const DIAGNOSES = makeRefRecords(ReferenceDataType.ICD10, ICD10_DIAGNOSES);
+const diagnosisRefData = makeRefRecords(ReferenceDataType.Diagnosis, DIAGNOSES);
 
-const dummyReferenceData: SyncRecord[] = [
-  ...VILLAGES,
-  ...DIAGNOSES,
-]
-  .map(data => ({
-    data,
-    recordType: 'referenceData',
-  }));
+const dummyReferenceData: SyncRecord[] = [...villageRefData, ...diagnosisRefData].map(data => ({
+  data,
+  recordType: 'referenceData',
+}));
 
-export const dummyReferenceRecords = [
-  ...dummyPatientRecords,
-  ...dummyReferenceData,
-].sort(sortByModified);
+export const dummyReferenceRecords = [...dummyPatientRecords, ...dummyReferenceData].sort(
+  sortByModified,
+);
