+++ packages/shared/src/demoData/index.js
@@ -1,20 +1,20 @@
 export * from './aefiSurvey';
 export * from './allergies';
 export * from './carePlan';
+export * from './createMockReportingSchemaAndRoles';
 export * from './departments';
 export * from './diagnoses';
 export * from './dispositions';
 export * from './drugs';
 export * from './facilities';
-export * from './icd10';
+export * from './insurers';
 export * from './labRequests';
 export * from './labTestTypes';
 export * from './locations';
 export * from './patients';
 export * from './procedureTypes';
-export * from './users';
-export * from './villages';
-export * from './utilities';
-export * from './createMockReportingSchemaAndRoles';
 export * from './settings';
-export * from './insurers';
+export * from './triageDiagnoses';
+export * from './users';
+export * from './utilities';
+export * from './villages';
